Improve signed file/signature file heuristic
- previously, did search for smaller filename before searching
  for longer filename. I.e., searched by removing extension
  first.
- now searching for an added extension first is more reliable
  in case a signed-encrypted file has been decrypted in the
  directory.

E.g., if a directory contains out, out.gpg, and out.gpg.sig, and
the user was verifying out.gpg, the old heuristic would see
out as the signed file and out.gpg as the signature file.

The new heuristic will match the longer name first, so it would
correctly process out.gpg as the data and out.gpg.sig as the
signature.